requirements.txt
django
djangorestframework
rembg
django-cors-headers
PyMuPDF
pdfkit
pytesseract